refactor: move avatar-lights component from registry to portfolio feature
Move avatar-lights component and update import paths to organize code
by feature rather than in shared registry structure
