+++ src/features/portfolio/types/user.ts
@@ -1,4 +1,4 @@
-import type { AvatarLightsVariants } from "@/registry/components/avatar-lights"
+import type { AvatarLightsVariants } from "@/features/portfolio/components/avatar-lights"
 
 export type User = {
   firstName: string
