--- src/registry/components/avatar-lights/index.ts
@@ -1,1 +1,0 @@
-export * from "./avatar-lights"
